+++ R/midnight-package.R
@@ -4,9 +4,9 @@
 ## usethis namespace: start
 #' @importFrom rlang '%||%'
 #' @importFrom rlang .data
-#' @importFrom parsnip new_model_spec
 #' @importFrom midr ggmid
 #' @importFrom Rcpp sourceCpp
 #' @useDynLib midnight, .registration = TRUE
 ## usethis namespace: end
 NULL
+