+++ configure.win.R
@@ -1,2 +1,0 @@
-#!/bin/sh
-echo "PKG_CXXFLAGS += -Wa,-mbig-obj" >> ./src/Makevars.win